move interface back to index.ts
